requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.34.0
tinydb==4.8.0
gunicorn==22.0.0
uvloop==0.19.0
httptools==0.6.1
python-dotenv==1.0.0
prisma==0.15.0
logging==0.4.9.6